abandon shared filter/sort; layout tweaks; fix sort bug
