+++ app/src/main/res/drawable/border.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <corners android:radius="1dp" />
+    <stroke
+        android:width="1dp"
+        android:color="@color/colorPrimary" />
+    <solid android:color="@color/colorTransparent" />
+</shape>
