+++ app/src/main/res/layout/cell_favorite_broadcast.xml
@@ -32,7 +32,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_marginTop="@dimen/timeslot_margin"
         android:layout_marginStart="@dimen/timeslot_margin"
-        android:layout_marginEnd="@dimen/spacing_small"
+        android:layout_marginEnd="@dimen/timeslot_margin"
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
 
@@ -80,12 +80,32 @@
             android:onClick="@{clickListener}"
             style="@style/TextAppearance.MaterialComponents.Headline6"
             app:layout_constraintStart_toEndOf="@id/broadcastImage"
-            app:layout_constraintEnd_toStartOf="@id/downloadDeleteIcon"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/sectionHeader"
             app:layout_constrainedWidth="true"
             app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintHorizontal_chainStyle="packed"
-            tools:text="Testing Overflow With a Long Name"/>
+            tools:text="Testing Overflow With a Very Long Name"/>
+
+        <TextView
+            android:id="@+id/showHost"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textColor="@android:color/white"
+            android:layout_marginStart="@dimen/timeslot_margin"
+            android:layout_marginEnd="@dimen/timeslot_margin"
+            android:maxLines="1"
+            android:ellipsize="end"
+            android:text="@{show.name}"
+            android:onClick="@{clickListener}"
+            style="@style/TextAppearance.MaterialComponents.Subtitle2"
+            app:layout_constraintStart_toEndOf="@id/broadcastImage"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/showName"
+            app:layout_constrainedWidth="true"
+            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintHorizontal_chainStyle="packed"
+            tools:text="DJ Stingray"/>
 
         <TextView
             android:id="@+id/date"
@@ -100,7 +120,7 @@
             localDate="@{broadcast.date}"
             style="@style/TextAppearance.MaterialComponents.Subtitle2"
             app:layout_constraintStart_toStartOf="@id/showName"
-            app:layout_constraintTop_toBottomOf="@id/showName"
+            app:layout_constraintTop_toBottomOf="@id/showHost"
             app:layout_constrainedWidth="true"
             app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintHorizontal_chainStyle="packed"
